- extend run id to 32 bits - use auto-increment setting for run id (Configuration->General Setup) - set run id via CARESS interface - make project files more readable - TODO: test directory has missing source files
git-svn-id: https://trac.frm2.tum.de/qmesydaq/trunk@752 71953026-4e7b-de11-b703-00e01807f32f

+++ lib/mcpd8.h
@@ -246,7 +246,7 @@ public:
 	//! \return the current time of the MCPD in msec
 	quint64 time(void) {return m_timemsec;}
 
-	bool setRunId(quint16 runid);
+	bool setRunId(quint32 runid);
 
 	/**
 	 * gets the current set run ID
@@ -254,7 +254,7 @@ public:
 	 * \return run ID
 	 * \see setRunId
 	 */
-	quint16 getRunId(void) {return m_runId;}
+	quint32 getRunId(void) {return m_runId;}
 
 	/**
 	 * gets the address of the MCPD module
@@ -393,7 +393,7 @@ private:
 	QTimer 		*m_commTimer;
 
 	//! current run ID
-	quint16 	m_runId;
+	quint32 	m_runId;
 
 	//! DAQ started?
 	bool		m_daq;
